support old carafe version and new carafe version

+++ modules/carafe.nf
@@ -50,7 +50,9 @@ process CARAFE {
             ${carafe_params} \\
         > >(tee "carafe.stdout") 2> >(tee "carafe.stderr" >&2)
 
-        mv -v SkylineAI_spectral_library.tsv carafe_spectral_library.tsv
+        # move SkylineAI_spectral_library.tsv (legacy compatibility)
+        [ -e SkylineAI_spectral_library.tsv ] && mv -v SkylineAI_spectral_library.tsv carafe_spectral_library.tsv
+
         echo "carafe_version=\$CARAFE_VERSION" > carafe_version.txt
         """
